# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: INDO DUTA TECH - Premium Management System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: units
    description: Laptop unit inventory management
  - name: dashboard
    description: Dashboard and analytics
  - name: settings
    description: Store settings and profile
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /units:
    get:
      operationId: listUnits
      tags: [units]
      summary: List all units
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PROSES, READY, TERJUAL]
          required: false
          description: Filter by unit status
      responses:
        "200":
          description: List of units
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Unit"
    post:
      operationId: createUnit
      tags: [units]
      summary: Create a new unit (Beli)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitInput"
      responses:
        "201":
          description: Unit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"

  /units/{id}:
    get:
      operationId: getUnit
      tags: [units]
      summary: Get a single unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Unit detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUnit
      tags: [units]
      summary: Update a unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitUpdate"
      responses:
        "200":
          description: Unit updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteUnit
      tags: [units]
      summary: Delete a unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /units/{id}/qc:
    post:
      operationId: completeQc
      tags: [units]
      summary: Complete QC inspection and mark unit as READY
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QcInput"
      responses:
        "200":
          description: Unit updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /units/{id}/jual:
    post:
      operationId: markSold
      tags: [units]
      summary: Mark unit as sold (TERJUAL)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JualInput"
      responses:
        "200":
          description: Unit sold
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /units/{id}/caption:
    get:
      operationId: getUnitCaption
      tags: [units]
      summary: Get marketing caption for a unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Caption text
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaptionResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /units/{id}/invoice:
    get:
      operationId: getUnitInvoice
      tags: [units]
      summary: Get invoice data for a sold unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Invoice document data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceDocument"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /units/{id}/kuitansi:
    get:
      operationId: getUnitKuitansi
      tags: [units]
      summary: Get kuitansi (receipt) data for a sold unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Kuitansi document data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KuitansiDocument"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get store settings
      responses:
        "200":
          description: Store settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreSettings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update store settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreSettingsInput"
      responses:
        "200":
          description: Updated store settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoreSettings"

components:
  schemas:
    HealthStatus:
      type: object
      required:
        - status
      properties:
        status:
          type: string

    UnitStatus:
      type: string
      enum: [PROSES, READY, TERJUAL]

    Unit:
      type: object
      required:
        - id
        - tipe
        - spek
        - minus
        - kelengkapan
        - hargaBeli
        - biayaQc
        - status
        - createdAt
      properties:
        id:
          type: number
        tipe:
          type: string
        spek:
          type: string
        minus:
          type: string
        kelengkapan:
          type: string
        hargaBeli:
          type: number
        biayaQc:
          type: number
        baterai:
          type: number
          nullable: true
        appTambahan:
          type: string
          nullable: true
        fisik:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/UnitStatus"
        hargaJual:
          type: number
          nullable: true
        tanggalJual:
          type: string
          format: date-time
          nullable: true
        namaPembeli:
          type: string
          nullable: true
        nomorPembeli:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time

    UnitInput:
      type: object
      required:
        - tipe
        - spek
        - minus
        - kelengkapan
        - hargaBeli
      properties:
        tipe:
          type: string
        spek:
          type: string
        minus:
          type: string
        kelengkapan:
          type: string
        hargaBeli:
          type: number

    UnitUpdate:
      type: object
      properties:
        tipe:
          type: string
        spek:
          type: string
        minus:
          type: string
        kelengkapan:
          type: string
        hargaBeli:
          type: number

    QcInput:
      type: object
      required:
        - fisik
        - baterai
      properties:
        fisik:
          type: string
        baterai:
          type: number
        appTambahan:
          type: string
          nullable: true
          description: Optional extra apps installed during QC
        checklistItems:
          type: array
          description: Detailed checklist records grouped by professional QC categories
          items:
            type: object
            required:
              - category
              - itemKey
              - itemLabel
              - status
            properties:
              category:
                type: string
                enum: [PHYSICAL, DISPLAY, HARDWARE_IO, PERFORMANCE_SOFTWARE]
              itemKey:
                type: string
              itemLabel:
                type: string
              status:
                type: string
                enum: [PASS, FAIL, N/A]
              sparepartId:
                type: number
                nullable: true
                description: Required when status is FAIL
              notes:
                type: string
                nullable: true

    JualInput:
      type: object
      required:
        - hargaJual
        - namaPembeli
        - nomorPembeli
      properties:
        hargaJual:
          type: number
        namaPembeli:
          type: string
        nomorPembeli:
          type: string

    CaptionResponse:
      type: object
      required:
        - caption
      properties:
        caption:
          type: string

    InvoiceDocument:
      type: object
      required:
        - nomorInvoice
        - tanggal
        - unit
        - store
        - subtotal
        - total
      properties:
        nomorInvoice:
          type: string
        tanggal:
          type: string
          format: date-time
        unit:
          $ref: "#/components/schemas/Unit"
        store:
          $ref: "#/components/schemas/StoreSettings"
        subtotal:
          type: number
        total:
          type: number

    KuitansiDocument:
      type: object
      required:
        - nomorKuitansi
        - tanggal
        - unit
        - store
        - jumlah
        - terbilang
      properties:
        nomorKuitansi:
          type: string
        tanggal:
          type: string
          format: date-time
        unit:
          $ref: "#/components/schemas/Unit"
        store:
          $ref: "#/components/schemas/StoreSettings"
        jumlah:
          type: number
        terbilang:
          type: string

    StoreSettings:
      type: object
      required:
        - id
        - namaToko
        - tagline
        - alamat
        - telepon
        - email
        - instagram
        - facebook
        - whatsapp
        - tiktok
        - website
        - logo
        - updatedAt
      properties:
        id:
          type: number
        namaToko:
          type: string
        tagline:
          type: string
        alamat:
          type: string
        telepon:
          type: string
        email:
          type: string
        instagram:
          type: string
        facebook:
          type: string
        whatsapp:
          type: string
        tiktok:
          type: string
        website:
          type: string
        logo:
          type: string
          description: Base64 data URL of store logo
        updatedAt:
          type: string
          format: date-time

    StoreSettingsInput:
      type: object
      properties:
        namaToko:
          type: string
        tagline:
          type: string
        alamat:
          type: string
        telepon:
          type: string
        email:
          type: string
        instagram:
          type: string
        facebook:
          type: string
        whatsapp:
          type: string
        tiktok:
          type: string
        website:
          type: string
        logo:
          type: string

    DashboardStats:
      type: object
      required:
        - totalModal
        - totalModalTerjual
        - totalUnitProses
        - totalUnitReady
        - totalUnitTerjual
        - estimasiNilaiJual
        - potensiProfit
        - realisasiProfit
        - recentUnits
      properties:
        totalModal:
          type: number
          description: Total capital invested in READY units (beli + QC)
        totalModalTerjual:
          type: number
          description: Total capital invested in TERJUAL units (beli + QC) — denominator for ROA
        totalUnitProses:
          type: number
          description: Units currently in QC process
        totalUnitReady:
          type: number
          description: Units ready for sale
        totalUnitTerjual:
          type: number
          description: Units sold
        estimasiNilaiJual:
          type: number
          description: Estimated total sell value of READY units
        potensiProfit:
          type: number
          description: Potential profit from READY units
        realisasiProfit:
          type: number
          description: Realized profit from TERJUAL units
        recentUnits:
          type: array
          description: Last 5 units added
          items:
            $ref: "#/components/schemas/Unit"

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string