# lib/api-zod/src/generated/types/qcInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * INDO DUTA TECH - Premium Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { QcInputChecklistItemsItem } from './qcInputChecklistItemsItem';

export interface QcInput {
  fisik: string;
  baterai: number;
  /** Optional extra apps installed during QC */
  appTambahan?: string | null;
  /** Detailed checklist records grouped by professional QC categories */
  checklistItems?: QcInputChecklistItemsItem[];
}